Solve UB and race condition through more raw ptrs
Removes pointers to the nodes and replaces them by storing everythin
needed in tasks incl. pointers to the dyn Gen.

Miri complained about a race condition so I changed the number of
necessary generation changes to 2. This should be a temporary solution
as 1 generation should be enough.

To make sure the inputs buffers of the Graph doesn't get freed, they are
stored in the GraphGen as well.